+++ src/sections/trade/TradePage.styled.ts
